done setup for language change
