+++ messages/hi.json
@@ -1,0 +1,5 @@
+{
+  "Index": {
+    "title": "मानक हिन्दी"
+  }
+}
