+++ messages/pu.json
@@ -1,0 +1,5 @@
+{
+  "Index": {
+    "title": "ਪੰਜਾਬੀ"
+  }
+}
