+++ src/app/[locale]/home/categories/index.tsx
