+++ src/app/[locale]/home/categories/style.sass
