+++ src/app/[locale]/home/collections/index.tsx
