+++ src/app/[locale]/home/deals/deals.sass
