+++ src/app/[locale]/home/deals/index.tsx
