+++ src/app/[locale]/home/khalsa/index.tsx
