+++ src/app/[locale]/home/khalsa/khalsa.sass
