+++ src/app/[locale]/home/products/index.tsx
