+++ src/app/[locale]/home/slider/index.tsx
