+++ src/app/[locale]/home/slider/slider.sass
