+++ src/app/[locale]/home/slider/slides.tsx
