+++ src/app/[locale]/home/zone/index.tsx
