+++ src/app/[locale]/home/zone/style.sass
