+++ src/app/[locale]/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next"
 import { Outfit } from "next/font/google"
-import '../shared/styles/main.sass'
+// import '../shared/styles/main.sass'
 import Header from "@/shared/components/header"
 import Footer from "@/shared/components/footer"
 
@@ -14,13 +14,15 @@ export const metadata: Metadata = {
   description: "Get new styles of T-shirts, Gatras and Bags"
 }
 
-export default function RootLayout({
-  children
+export default function LocaleLayout({
+  children,
+  params: { locale }
 }: Readonly<{
   children: React.ReactNode
+  params: { locale: string };
 }>): React.ReactNode {
   return (
-    <html lang="en">
+    <html lang={locale}>
       <body className={outfit.className}>
         <Header />
         <main>
