+++ src/app/[locale]/shop/page.tsx
