+++ src/app/[locale]/test/page.tsx
