+++ src/app/[locale]/wishlist/page.tsx
